fix trt plugin clone and initialize bugs in TRT7.1+ (#30709)
* fix trt plugin clone and initialize bugs

* fix unit test error

* enable trt in ci py3

* update unittest timeout
