+++ python/paddle/fluid/tests/unittests/ir/inference/CMakeLists.txt
@@ -30,4 +30,6 @@ foreach(target ${TEST_INFERENCE_IR_PASSES})
 endforeach()
 if(WITH_GPU AND TENSORRT_FOUND)
 set_tests_properties(test_trt_subgraph_pass PROPERTIES TIMEOUT 120)
+set_tests_properties(test_trt_activation_pass PROPERTIES TIMEOUT 120)
+set_tests_properties(test_trt_conv_pass PROPERTIES TIMEOUT 120)
 endif()
